this is a pretty studly commit...read on
So, I was running into a bug when generating
fading arcs concurrently, namely, we want
the arcs to be dropped from the scene after
they dissolve, or fade away.  The problem is,
we can't concurrently update the structure
of the scene (namely parent/child relations)
without causing big problems.  Specifically,
we'd get an error about having an arraylist
index out of range (the dropped child caused
an off-by-one error during rendering) when
we allowed a drop to happen mid-frame.
Solution is to queue up disruptive tasks
like drops, and have piccolo's rendering
thread run them on its schedule, during
frame creation.  So, we need a "hook"
to schedule stuff with piccolo, which is
in java... The facilities piccolo provides
to do this are either via swing event
listeners, or piccolo "activities".
Turns out, we can create custom
activities by setting a delegate
via a reified interface (which clojure
handles gracefully), rather than the
inheritance abstraction "most" of
piccolo relies on.  So, similar to
how we did node-meta, I wrap the
ability to create activities (clojure
thunks), and install a "Service" that
runs as a piccolo activity every 16 ms.
It maintains a schedule, a concurrent queue
that can be appended to safely by multiple
threads (or goroutines).  It simply flushes
the buffer as it's step function, polling
the queue and evaluating each thunk.  For
convenience, I added a macro, schedule!,
to help inline these sorts of tasks
transparently.  Coolest part of all: it
worked on the first try.